src/backend/recommender/requirements.txt
flask
flask-cors
torch --extra-index-url https://download.pytorch.org/whl/cpu
torchvision
transformers
faiss-cpu
Pillow
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
firebase-admin
google-cloud-storage
google-cloud-firestore
requests
